gems_deals/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
gunicorn==20.1.0
python-dotenv==1.0.0
